stack upgrade test: pull lifecycle image
The paketo bionic builder is now dead, and so its lifecycle image version is not in lockstep with jammy builder anymore.

Fixes error "failed to build: fetching lifecycle image: image 'buildpacksio/lifecycle:0.16.5' does not exist on the daemon"

(Note: you see this error if you flip the order of builders in
integration.json. Currently the error is hidden because this test on
bionic is run after tests on jammy)